intégration stripe checkout + portail

+++ app/api/board/route.js
@@ -25,7 +25,15 @@ export async function POST(req) {
     }
     await connectMongo();
 
+
     const user = await User.findById(session.user.id);
+
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "Veuillez vous inscrire pour obtenir cette incroyable fonctionnalité 🦄"},
+        { status: 403 }
+      );
+    }
 
     const board = await Board.create({
       userId: user._id,
@@ -64,12 +72,23 @@ export async function DELETE(req) {
       );
     }
 
+
+
+    const user = await User.findById(session?.user?.id);
+
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "Veuillez vous inscrire pour obtenir cette incroyable fonctionnalité 🦄"},
+        { status: 403 }
+      );
+    }
+
     await Board.deleteOne ({
       _id: boardId,
       userId: session?.user?.id,
     });
 
-    const user = await User.findById(session?.user?.id);
+
     user.boards = user.boards.filter((id) => id.toString() !== boardId);
     await user.save();
     return NextResponse.json({});
